Practice Branch for understanding delete - working

+++ src/main/java/org/launchcode/liftoffrecipetracker/controllers/CategoryController.java
@@ -2,6 +2,7 @@ package org.launchcode.liftoffrecipetracker.controllers;
 
 import org.launchcode.liftoffrecipetracker.data.CategoryRepository;
 import org.launchcode.liftoffrecipetracker.models.Category;
+import org.launchcode.liftoffrecipetracker.models.Recipe;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -9,6 +10,8 @@ import org.springframework.validation.Errors;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
+import java.util.List;
+import java.util.Optional;
 
 
 @Controller
@@ -16,8 +19,8 @@ import javax.validation.Valid;
 @RequestMapping("categories")
 public class CategoryController {
 
-@Autowired
-private CategoryRepository categoryRepository;
+    @Autowired
+    private CategoryRepository categoryRepository;
 
     @GetMapping
     public String displayAllCategories(Model model) {
@@ -38,7 +41,7 @@ private CategoryRepository categoryRepository;
 
     @PostMapping("create")
     public String processCreateCategoryForm(@Valid @ModelAttribute Category category,
-                                                 Errors errors, Model model) {
+                                            Errors errors, Model model) {
 
         if (errors.hasErrors()) {
             model.addAttribute("title", "Create Category");
@@ -53,23 +56,48 @@ private CategoryRepository categoryRepository;
     }
 
     @GetMapping("delete")
-    public String displayDeleteCategoryForm(Model model){
+    public String displayDeleteCategoryForm(Model model) {
         model.addAttribute("title", "Delete Category");
         model.addAttribute("categories", categoryRepository.findAll());
-        return"category/delete";
+        return "category/delete";
     }
 
     @PostMapping("delete")
-    public String processDeleteCategoryForm( @RequestParam(required = false)
-                                                   int[] categoryId){
+    public String processDeleteCategoryForm(@RequestParam(required = false)
+                                                    int[] categoryId) {
 
-        if(categoryId != null){
-            for(int id : categoryId){
-                categoryRepository.deleteById(id);
+        if (categoryId != null) {
+            for (int id : categoryId) {
+                Optional<Category> optionalCategory = categoryRepository.findById(id);
+                if (optionalCategory.isPresent()) {
+                    Category category = optionalCategory.get();
+                    List<Recipe> recipes = category.getRecipes();
+                    for (Recipe recipe : recipes) {
+                        recipe.removeCategory(category);
+                    }
+                    recipes.clear();
+                    categoryRepository.deleteById(id);
+                }
             }
         }
         return "redirect:";
     }
-
-
 }
+
+
+////update category
+//@GetMapping("create/{categoryId}")
+//public String displayUpdateCategoryForm(@RequestParam int categoryId, Model model) {
+//
+//        Optional<Category> result = categoryRepository.findById(categoryId);
+//
+//        if (result.isEmpty()) {
+//            model.addAttribute("title", "Invalid Category ID: " + categoryId);
+//        } else {
+//            Category category = result.get();
+//            model.addAttribute("title", "Category Details: " + category.getName());
+//            model.addAttribute("category", category);
+//        }
+//        return "create";
+//    }
+//}
